requirements.txt
pysmb==1.2.10
jinja2==3.1.3
# ipaddress is built-in for Python 3.7+, no need to install
python-docx==0.8.11
openpyxl==3.1.2
pdfplumber==0.10.3
python-magic==0.4.27
# requests is used for geo IP lookups (optional feature)
requests==2.31.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
mypy>=1.0.0